+++ Views/Subjects/index.php
@@ -1,0 +1,27 @@
+<link rel="stylesheet" href="../../src/FOIL-UI.css" type="text/css">
+<link rel="stylesheet" href="../../src/config.css" type="text/css">
+<link rel="stylesheet" href="../../src/position-text.css" type="text/css">
+<div class="FOIL-font-styles sans-serif">
+<div class="FOIL-navigation-bar-regular">
+<h1> Subjects </h1>
+<a href="../../index.php" class="FOIL-button-regular-no-shadow" style="background-color:rgb(165, 148, 103); color: black;">Back</a>
+</div>
+
+<div class="FOIL-layout-regular-no-shadow">
+        <center> <FOIL-font-size-60> Subjects </FOIL-font-size-60> </center>
+        <div class="space"></div>
+        <div class="space"></div>
+        <div class="space"></div>
+        <div class="space"></div>
+        <div class="space"></div>
+        <center> <FOIL-font-size-20> ... </FOIL-font-size-20> </center>
+        <div class="space"></div>
+        <div class="space"></div>
+        <div class="space"></div>
+        <center>
+        <a href="#" class="FOIL-button-regular-no-shadow"></a>
+        <a href="#" class="FOIL-button-regular-no-shadow"></a>
+        <a href="#" class="FOIL-button-regular-no-shadow"></a>
+        <a href="#" class="FOIL-button-regular-no-shadow"></a>
+        </center>
+</div>